Removed bluebird promise wrap in request lib (#9343)
refs #9178, refs #8988

With https://github.com/TryGhost/Ghost/commit/4c54f3bb9cdcc3fa6ee1186515322629be83cce7 we use Bluebird globally for Promises. Therefore, the request lib doesn't need to be wrapped in a bluebird Promise anymore.

This was originally done, so we can work with catch predicated in our image-size lib.

Updated the tests to proof, that the catch predicates work.

The tests fail, as soon as the Promise overwrite is commented out.