🐛 Updated stripe prices/products table population
refs https://github.com/TryGhost/Team/issues/591
refs https://github.com/TryGhost/Ghost/commit/a08883cf33d591570f99036d2df9a775f1231657

It's possible to have sites which still have customer subscriptions in their DB from old Stripe accounts, most likely added when we allowed Stripe Direct, as those subscriptions were not cleaned up. While populating stripe prices and products for existing subscriptions, we want to ignore these old subscriptions which are not part of current Stripe account instead of throwing error and halting the one off migration which we currently do.

The population script will re-run again as long as no new subscriptions are created between the last release and this, and there were no old invalid plans in DB, which is a very unlikely edge case so the data will eventually re-populate correctly.